m/z to double type

+++ Apache/applet/SearchPage/PackageRecData.java
@@ -20,7 +20,7 @@
  *
  * レコード情報格納 クラス
  *
- * ver 1.0.2 2008.12.05
+ * ver 1.0.3 2009.12.14
  *
  ******************************************************************************/
 
@@ -65,7 +65,7 @@ public class PackageRecData {
 	private int peakNum;
 	
 	/** m/z */
-	private float[] mz;
+	private double[] mz;
 	
 	/** 強度 */
 	private int[] intensity;
@@ -246,7 +246,7 @@ public class PackageRecData {
 	 * @param end マスレンジ(m/z)終了値
 	 * @return レコード内の指定されたマスレンジ(m/z)の間で最大の強度
 	 */
-	public int getMaxIntensity(float start, float end) {
+	public int getMaxIntensity(double start, double end) {
 		int max = 0;
 		for (int i=0; i<this.peakNum; i++) {
 			if (this.mz[i] > end) {
@@ -279,7 +279,7 @@ public class PackageRecData {
 	 * @param index インデックス
 	 * @return m/z
 	 */
-	public float getMz(int index) {
+	public double getMz(int index) {
 		return mz[index];
 	}
 
@@ -289,15 +289,15 @@ public class PackageRecData {
 	 * @param mz m/z
 	 */
 	public void setMz(int index, String mz) {
-		this.mz[index] = Float.parseFloat(mz);
+		this.mz[index] = Double.parseDouble(mz);
 	}
 	
 	/**
 	 * 最大m/zとプリカーサーの比較
 	 * @return 最大m/zとプリカーサーの大きい方
 	 */
-	public float compMaxMzPrecusor() {
-		float mzMax;
+	public double compMaxMzPrecusor() {
+		double mzMax;
 		if (mz == null || mz.length == 0) {
 			mzMax = 0f;
 		}
@@ -305,12 +305,12 @@ public class PackageRecData {
 			mzMax = mz[mz.length-1];
 		}
 		try {
-			Float.parseFloat(precursor);
+			Double.parseDouble(precursor);
 		} catch (Exception e) {
 			return mzMax;
 		}
 		
-		return Math.max(mzMax, Float.parseFloat(precursor));
+		return Math.max(mzMax, Double.parseDouble(precursor));
 	}
 	
 	/**
@@ -331,7 +331,7 @@ public class PackageRecData {
 		
 		int index = -1;
 		for (int i=0; i<peakNum; i++) {
-			if (this.mz[i] == Float.parseFloat(mz)) {
+			if (this.mz[i] == Double.parseDouble(mz)) {
 				index = i;
 				break;
 			}
@@ -361,7 +361,7 @@ public class PackageRecData {
 	public void setSelectPeak(String mz, boolean status) {
 		int index = -1;
 		for (int i=0; i<peakNum; i++) {
-			if (this.mz[i] == Float.parseFloat(mz)) {
+			if (this.mz[i] == Double.parseDouble(mz)) {
 				index = i;
 				break;
 			}
@@ -477,7 +477,7 @@ public class PackageRecData {
 	 */
 	public void setPeakNum(int peakNum) {
 		this.peakNum = peakNum;
-		this.mz = new float[peakNum];
+		this.mz = new double[peakNum];
 		this.intensity = new int[peakNum];
 		this.selectPeak = new boolean[peakNum];
 		this.peakColorType = new int[peakNum];
@@ -489,7 +489,7 @@ public class PackageRecData {
 	 * @param target m/z指定値
 	 * @return インデックス
 	 */
-	public int getIndex(float target) {
+	public int getIndex(double target) {
 		int index;
 		for (index=0; index<this.peakNum; index++) {
 			if (this.mz[index] >= target) {
